requirements.txt
autopep8==1.5.7
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
python-decouple==3.5
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.1
wincertstore==0.2
